let isMorphable(A,B) be true when A and B have a shared array property of unequal length. One test modified.  All other tests pass.
The justification is that most buyer/seller property arrays would be round robin
and could be of unequal length at this test.  During a morph the unequal length
arrays are explicitly expanded (except for buyerValues/sellerCosts) by
a function renamed to 'explicitlyExpandToFitNumberOfAgents'

An issue can arise when A.buyerValues.length!==B.buyerValues.length and also
with sellerCosts; morph may work but will discard elements. This could
be addressed in the future by filling in Ls/Hs or 0s/max Vs to make
the two buyerValues schedules or sellerCosts schedules of equal length.